+++ LethalLevelLoader/General/Patches.cs
@@ -545,14 +545,31 @@ if (AssetBundleLoader.noBundlesFound == true)
             LevelLoader.RefreshTimeOfDayMusic(LevelManager.CurrentExtendedLevel);
         }
 
-        [HarmonyPatch(typeof(StartOfRound), "StartGame"), HarmonyPrefix, HarmonyPriority(priority)]
-        internal static void StartOfRoundStartGame_Prefix()
+        [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.StartGame)), HarmonyTranspiler, HarmonyPriority(priority)]
+        internal static IEnumerable<CodeInstruction> StartOfRoundStartGame_Transpiler(IEnumerable<CodeInstruction> instructions)
+        {
+            return new CodeMatcher(instructions).MatchForward(useEnd: false,
+                new(OpCodes.Ldarg_0),
+                new(OpCodes.Call, AccessTools.Method(typeof(NetworkManager), "get_NetworkManager")),
+                new(OpCodes.Callvirt, AccessTools.Method(typeof(NetworkSceneManager), "get_SceneManager")),
+                new(OpCodes.Ldarg_0),
+                new(OpCodes.Ldfld, AccessTools.Field(typeof(StartOfRound), nameof(StartOfRound.currentLevel))),
+                new(OpCodes.Ldc_I4_1),
+                new(OpCodes.Callvirt, AccessTools.Method(typeof(NetworkSceneManager), nameof(NetworkSceneManager.LoadScene))),
+                new(OpCodes.Pop))
+            .Insert( // Insert call to select a random scene immediately before the current scene begins to load, and after generating the seed for the current round.
+                new(OpCodes.Ldarg_0),
+                new(OpCodes.Call, AccessTools.Method(typeof(Patches), nameof(PerformSceneSelection))))
+            .InstructionEnumeration();
+        }
+
+        private static void PerformSceneSelection()
         {
             ExtendedLevel extendedLevel = LevelManager.CurrentExtendedLevel;
             if (!IsServer || extendedLevel == null) return;
 
             extendedLevel.SelectableLevel.sceneName = string.Empty;
-            RoundManager.InitializeRandomNumberGenerators();
+            System.Random levelRandom = new(StartOfRound.randomMapSeed);
 
             int counter = 1;
             foreach (StringWithRarity sceneSelection in extendedLevel.SceneSelections)
@@ -561,10 +578,22 @@ if (AssetBundleLoader.noBundlesFound == true)
                 counter++;
             }
 
-            List<int> sceneSelections = extendedLevel.SceneSelections.Select(s => s.Rarity).ToList();
-            int selectedSceneIndex = RoundManager.GetRandomWeightedIndex(sceneSelections.ToArray(), RoundManager.LevelRandom);
+            int[] sceneSelections = extendedLevel.SceneSelections.Select(s => s.Rarity).ToArray();
+            int selectedSceneIndex = RoundManager.GetRandomWeightedIndex(sceneSelections, levelRandom);
             extendedLevel.SelectableLevel.sceneName = extendedLevel.SceneSelections[selectedSceneIndex].Name;
             DebugHelper.Log("Selected SceneName: " + extendedLevel.SelectableLevel.sceneName + " For ExtendedLevel: " + extendedLevel.NumberlessPlanetName, DebugType.Developer);
+        }
+
+        [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.SceneManager_OnLoadComplete1)), HarmonyPrefix, HarmonyPriority(priority)]
+        internal static void StartOfRoundOnLoadComplete_Prefix(string sceneName)
+        {
+            ExtendedLevel extendedLevel = LevelManager.CurrentExtendedLevel;
+            if (extendedLevel == null || extendedLevel.SelectableLevel.sceneName == sceneName) return;
+
+            if (extendedLevel.SceneSelections.Select(scene => sceneName).Contains(sceneName)) // Check if a valid scene loaded.
+                extendedLevel.SelectableLevel.sceneName = sceneName; // Update current level's scene name, so the round can end properly.
+            else
+                DebugHelper.LogError($"Critical Failure! Scene '{sceneName}' has no selection entry for ExtendedLevel {extendedLevel.NumberlessPlanetName}!", DebugType.User);
         }
 
         [HarmonyPatch(typeof(DungeonGenerator), "Generate"), HarmonyPrefix, HarmonyPriority(priority)]
@@ -768,11 +797,11 @@ if (AssetBundleLoader.noBundlesFound == true)
 
         private static bool CheckLever(InteractTrigger trigger)
         {
-            trigger.interactable = NetworkBundleManager.AllowedToLoadLevel && !StartOfRound.Instance.travellingToNewLevel;
+            trigger.interactable = NetworkBundleManager.AllowedToLoadLevel && !StartOfRound.travellingToNewLevel;
 
             if (!trigger.interactable)
             {
-                trigger.disabledHoverTip = StartOfRound.Instance.travellingToNewLevel ? routingText : disabledText;
+                trigger.disabledHoverTip = StartOfRound.travellingToNewLevel ? routingText : disabledText;
 
                 return false;
             }